pyproject.toml
[project]
name = "wikipedia-neo4j-demo"
version = "0.1.0"
description = "Minimal Wikipedia -> Neo4j GraphRAG demo"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "neo4j>=5.23.0",
  "wikipedia>=1.4.0",
  "pydantic-settings>=2.4.0",
  "google-genai>=1.32.0",
  "datasets>=3.0.0"
]

[dependency-groups]
dev = [
  "mkdocs>=1.6.0,<2.0.0",
  "mkdocs-material>=9.5.0,<10.0.0",
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.12.0",
  "mypy>=1.11.0"
]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = true
exclude = ["site/"]

[tool.pytest.ini_options]
addopts = "-q --cov=src --cov-report=term-missing --cov-fail-under=75"
testpaths = ["tests"]
